refactor(neon_talk): Move logic for rich object file previews to a separate widget
